Move features to registrations.
